creador_de_qr/requirements.txt
Pillow==11.3.0
qrcode==8.2